+++ cli/srecli/group/users.py
@@ -46,19 +46,25 @@ def create(ctx: ContextObj, username, password, role_name, customer_id):
         username=username,
         password=password,
         customer_id=customer_id,
-        role_name=role_name
+        role_name=role_name,
     )
 
 
 @users.command(cls=ViewCommand, name='update')
 @click.option('--username', required=True, type=str,
-              help='Username to create user')
+              help='Username to update user')
 @click.option('--password', '-p', type=str, help='New user password')
 @click.option('--role_name', '-rn', type=str,
               help='Role to assign to this user. '
                    'It should exist inside the customer')
 @cli_response()
-def update(ctx: ContextObj, username, customer_id, password, role_name):
+def update(
+    ctx: ContextObj,
+    username: str,
+    customer_id: str,
+    password: str | None,
+    role_name: str | None,
+):
     """
     Updates some user's attributes
     """
@@ -66,13 +72,13 @@ def update(ctx: ContextObj, username, customer_id, password, role_name):
         username=username,
         customer_id=customer_id,
         password=password,
-        role_name=role_name
+        role_name=role_name,
     )
 
 
 @users.command(cls=ViewCommand, name='delete')
 @click.option('--username', required=True, type=str,
-              help='Username to create user')
+              help='Username to delete user')
 @cli_response()
 def delete(ctx: ContextObj, username, customer_id):
     """